backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-docx
pypdf
celery
redis
httpx
python-dotenv
alembic
psycopg2-binary
aiofiles
tenacity
